+++ actions/browsing/README.md
@@ -1,0 +1,76 @@
+# Browsing
+
+Package provides base actions needed to get information from websites and interact with them.
+Currently supports file downloads, filling form elements and getting website content.
+
+Website content retrieved:
+
+    - text content of the page
+    - elements ('a', 'input', 'button', 'select', 'textarea' and 'table')
+
+## Prompts
+
+```
+Get top links from sema4.ai website.
+```
+
+> Here are the top links from the Sema4.ai website:
+>
+> 1. Product
+> 2. About us
+> 3. Blog
+> 4. Contact us
+> 5. Read our blog
+> 6. Build an AI Action
+>    ...
+
+```
+Get main content from bbc.com
+```
+
+> Here is the main content from the BBC website:
+>
+> Top News Stories
+>
+> 1. Exploding batteries spark deadly S Korea factory fire
+>    At least 22 people have been confirmed dead while eight others were injured, local authorities say.
+>    Read more
+>
+> 2. Netanyahu: Israeli forces will move to Lebanon border as Rafah winds down
+>    Israel’s PM says the end of the Gaza war's "intense phase” will allow forces to confront Hezbollah.
+>    Read more
+>
+> 3. Russia blames US for Crimea deaths and vows response
+>    Russia says US-made missiles programmed by American specialists were used in an attack which killed four.
+>    Read more
+>    ...
+>
+> Featured Articles
+>
+> 1. Magic Johnson on the 1992 Olympic Dream Team
+>    The American basketball great and renowned businessman opened up to the BBC's Katty Kay about family, his health, and life's greatest moments.
+>    Read more
+> 2. 'I only learnt to know her through her diary'
+>    On 25 June 1947, Anne Frank's diary was first published. In this exclusive archive clip, her father tells the BBC about his decision to make her words public.
+>    Read more
+>
+> Sports
+>
+> 1. UEFA Euro 2024
+>    Live updates
+>    Euro 2024 latest: Gallagher likely to start against Slovenia and Shaw in full training.
+>    What now for Clarke as Scotland's ongoing deterioration continues?
+>    Scotland head coach Steve Clarke must realize that more of the same is not good enough after his side's Euro 2024 exit.
+>    Read more
+>    ...
+>
+> Additional News
+>
+> 1. How exam scandals threaten the future of India's young people
+>    Paper leaks have plunged India's exams into chaos, jeopardizing the careers of millions of aspirants.
+>    Read more
+>
+> 2. Parasites and old clothes in North Korea trash balloons
+>    More than 1,500 balloons have been sent into South Korea from the North since May, officials say.
+>    Read more
+>    ....
