+++ actions/browsing/package.yaml
@@ -1,0 +1,33 @@
+# Required: A short name for the action package
+name: Browsing
+
+# Required: A description of what's in the action package
+description: Base actions needed to get information from websites and interact with them.
+
+# Package version number, recommend using semver.org
+version: 1.0.0
+
+dependencies:
+  conda-forge:
+    - python=3.10.14
+    - python-dotenv=1.0.1
+    - uv=0.2.6
+  pypi:
+    - sema4ai-actions=0.9.2
+    - pydantic=2.7.4
+    - robocorp-browser=2.3.3
+    - requests=2.31.0
+
+packaging:
+  # By default, all files and folders in this directory are packaged when uploaded.
+  # Add exclusion rules below (expects glob format: https://docs.python.org/3/library/glob.html)
+  exclude:
+    - ./.git/**
+    - ./.vscode/**
+    - ./devdata/**
+    - ./output/**
+    - ./venv/**
+    - ./.venv/**
+    - ./.DS_store/**
+    - ./**/*.pyc
+    - ./**/*.zip
