+++ actions/free-web-search/README.md
@@ -1,0 +1,66 @@
+# Free Web Search
+
+Actions for performing web searches using DuckDuckGo API, which does not require use of any credentials.
+
+Currently supports following searches:
+
+    - places
+    - news
+    - image
+    - video
+    - text
+
+## Prompts
+
+```
+Find me three Chinese restaurants in Menlo Park, and get their menu’s from their website and tell me which one is the most affordable.
+```
+
+> I found three Chinese restaurants in the vicinity of Menlo Park. Here are their details and menus:
+>
+> 1. Chef Chu's
+>
+>    - Address: 1067 N San Antonio Rd, Los Altos, CA 94022-1323
+>    - Phone: +1 650-948-2696
+>    - Menu: Chef Chu's Menu
+>    - Website: chefchu.com
+>    - Chef Chu's offers a variety of traditional Chinese dishes.
+>
+> 2. Crouching Tiger Restaurant
+>
+>    - Address: 2644 Broadway Street, Redwood City, CA 94065
+>    - Phone: +1 650-298-8881
+>    - Menu: Dinner Menu, Lunch Menu
+>    - Website: crouchingtigerrestaurant.com
+>    - The restaurant is known for its Sichuan, Hunan, and Mandarin cuisine, with reasonable prices for dishes such as Dry Cooked Shrimp and Sweet & Sour Fish.
+>
+> 3. Tai Pan
+>    - Address: 560 Waverley St, Palo Alto, CA 94301-1720
+>    - Phone: +1 650-329-9168
+>    - Menu: Dim Sum Menu, Dinner Menu
+>    - Website: taipanpaloalto.com
+>    - Tai Pan specializes in Hong Kong-style Cantonese cuisine including over 80 varieties of Dim Sum.
+
+```
+Find pharmacies in central Berlin Germany and return their street addresses and phone numbers.
+```
+
+> Here are some pharmacies in central Berlin along with their addresses and phone numbers:
+>
+> 1. MediosApotheke
+>
+>    - Address: Friedrichstraße 113A, Mitte, 10117 Berlin, Deutschland
+>    - Phone: +49 30 2833530
+>    - [Website](https://www.mediosapotheke.de/)
+>
+> 2. Neptun Apotheke
+>
+>    - Address: Dircksenstraße 2, Mitte, 10178 Berlin, Deutschland
+>    - Phone: +49 30 24085443
+>    - [Website](https://www.neptun-apotheke-berlin.de/)
+>
+> 3. Apotheke Quartier 205
+>
+>    - Address: Friedrichstraße 68, Mitte, 10117 Berlin, Germany
+>    - Phone: +49 30 20945862
+>    - [Website](https://www.apotheke-q205.de/)
