+++ actions/free-web-search/devdata/input_web_search_images.json
@@ -1,0 +1,3 @@
+{
+    "keywords": "butterfly"
+}
