+++ actions/free-web-search/devdata/input_web_search_videos.json
@@ -1,0 +1,3 @@
+{
+    "keywords": "AI and Robots"
+}
