+++ actions/google-search/.gitignore
@@ -1,0 +1,23 @@
+*.pkl
+*.pyc
+*.zip
+*/.ipynb_checkpoints/*
+*/.virtual_documents/*
+*/work-items-out/*
+.DS_Store
+.env
+.idea/
+.ipynb_checkpoints/
+.project
+.pydevproject
+.rpa/
+.use
+.venv/
+.virtual_documents/
+.vscode
+.~lock*
+metadata.json
+output/
+temp/
+testrun/*
+venv/
