+++ actions/google-search/devdata/.env.template
@@ -1,0 +1,2 @@
+GOOGLE_SEARCH_API_KEY=<your_google_search_api_key>
+GOOGLE_SEARCH_CONTEXT=<your_google_search_engine_context_id>
