+++ actions/google-search/docs/images/6_customize_engine.png
