+++ actions/google-search/package.yaml
@@ -1,8 +1,8 @@
 # Required: A short name for the action package
-name: Search and Browse
+name: Google Search
 
 # Required: A description of what's in the action package
-description: Base actions needed to get information from websites and interact with them.
+description: Google search action requiring API and context keys from Google.
 
 # Package version number, recommend using semver.org
 version: 1.0.0
@@ -15,9 +15,7 @@ dependencies:
   pypi:
     - sema4ai-actions=0.9.2
     - pydantic=2.7.4
-    - robocorp-browser=2.3.3
     - requests=2.31.0
-    - duckduckgo-search=5.3.0
 
 packaging:
   # By default, all files and folders in this directory are packaged when uploaded.
