--- actions/search-and-browse/devdata/.env.template
@@ -1,2 +1,0 @@
-GOOGLE_SEARCH_API_KEY=<your_google_search_api_key>
-GOOGLE_SEARCH_CONTEXT=<your_google_search_engine_context_id>
